unix: Fix missing REPL output
MicroPython uses a custom readline implementation which echoes back all
characters in raw mode, hereby making it unsuitable for use with standard
REPLWrapper functionality (i.e. split notebook input in lines, feed all
lines into microPython). The previous solution was to compile all input
at once then execute it. Not a bad idea, where it not that by doing so
the result of expressions in the input do not get printed out and hence
don't get sent back to the notebook, which sort of defeats the purpose.

Fix this by reverting to use standard REPLWrapper functionality in
combination with a MicroPython build using standard readline.
The drawback is this requires a patch for MicroPython since it partially
broke readline support a while ago - see the README.
A future fix could be to adapt the MicroPython readline implementation
to behave more like a standard one.